+++ CHANGELOG.md
@@ -52,7 +52,7 @@
   [#506](https://github.com/Kong/gateway-operator/pull/506)
 - Add `KongConsumerGroup` reconciler for Konnect control planes.
   [#510](https://github.com/Kong/gateway-operator/pull/510)
-- Add `KongCACertificate` reconciler for Konnect control planes.
+- Add `KongCACertificate` reconciler for Konnect CA certificates.
   [#626](https://github.com/Kong/gateway-operator/pull/626)
 - Added command line flags to configure the certificate generator job's images.
   [#516](https://github.com/Kong/gateway-operator/pull/516)